finish sections 1, 2 and 3
Finally! Got it to work. Removed 'justify-content: center' from
the main content container which was causing the problem.

Now have to finish section 4 and I am done.

Wonder if I should change the header to not split as much, but that
is an extreme case when I have zoomed out so much.